fix: don't set OAuth endpoint if authority attribute not available
If MicrosoftAppCredentials doesn't support the 'authority' attribute,
don't try to set the OAuth endpoint (which causes the SDK to treat it
as an authority and fail). Instead, let the adapter use its default
credentials, which should work for single-tenant bots if the Azure
App Registration is configured correctly as 'Single tenant'.